requirements.txt
Flask>=2.2,<2.4
feedparser>=6.0
requests>=2.0
python-dotenv>=1.0
